первый релиз!!!!

+++ app.py
@@ -4,6 +4,7 @@ from werkzeug.utils import redirect
 import sqlite3
 
 app = Flask(__name__)
+ADMIN_PASSWORD = "okyrvamatka"
 
 def init_db():
     with sqlite3.connect("games.db") as conn:
@@ -45,6 +46,12 @@ def form_add_game():
 
 @app.route("/add_games", methods=['POST'])
 def add_game():
+    password = request.form.get("password")
+
+
+    if password != ADMIN_PASSWORD:
+        return """<script>alert("Неверный пароль!"); window.history.back();</script>""", 403
+
     title = request.form["title"]
     description_game = request.form["description"]
     mi_number_player = request.form["min_number_player"]
@@ -57,7 +64,14 @@ def add_game():
         conn.execute("""
             INSERT INTO games (title, description, min_players, max_players, min_play_time, max_play_time, UPDT)
             VALUES (?, ?, ?, ?, ?, ?, ?)
-        """, (title, description_game, mi_number_player, ma_number_player, mi_playing_time, ma_playing_time, UPDT))
+        """, (
+            title,
+            description_game,
+            mi_number_player,
+            ma_number_player,
+            mi_playing_time,
+            ma_playing_time,
+            UPDT))
     return redirect("/add")
 
 
@@ -66,6 +80,10 @@ def add_game():
 
 @app.route("/update_game", methods = ['POST'])
 def up_gam():
+
+    password = request.form.get("password")
+    if password != ADMIN_PASSWORD:
+        return """<script>alert("Неверный пароль!"); window.history.back();</script>""", 403
     with sqlite3.connect("games.db") as conn:
         cursor = conn.cursor()
         number_game = request.form["num_gam"]
@@ -80,10 +98,36 @@ def up_gam():
 
 @app.route("/save_update", methods = ['POST'])
 def s_u():
+    number_game = request.form["num_gam"]
+    title = request.form["title"]
+    description_game = request.form["description"]
+    mi_number_player = request.form["min_number_player"]
+    ma_number_player = request.form["max_number_player"]
+    mi_playing_time = request.form['min_playing_time']
+    ma_playing_time = request.form['max_playing_time']
+    UPDT = 1
     with sqlite3.connect("games.db") as conn:
         cursor = conn.cursor()
-        number_game = request.form["num_gam"]
-        cursor.execute(f"UPDATE games SET UPDT = 1 WHERE id = {number_game}")
+        conn.execute("""
+        UPDATE games SET 
+        title = ?, 
+        description = ?, 
+        min_players = ?, 
+        max_players = ?, 
+        min_play_time = ?, 
+        max_play_time = ?,
+        UPDT = 1 
+        WHERE id = ?
+        """,
+(
+        title,
+        description_game,
+        mi_number_player,
+        ma_number_player,
+        mi_playing_time,
+        ma_playing_time,
+        number_game
+        ))
         cursor.execute("SELECT id, title, description, min_players, max_players, min_play_time, max_play_time, UPDT FROM games")
         games = cursor.fetchall()
     return render_template("games.html", games=games)
@@ -95,6 +139,10 @@ def s_u():
 
 @app.route("/delete_game", methods = ['POST'])
 def del_gam():
+    password = request.form.get("password")
+    if password != ADMIN_PASSWORD:
+        return """<script>alert("Неверный пароль!"); window.history.back();</script>""", 403
+
     number_game = request.form["num_gam"]
     with sqlite3.connect("games.db") as conn:
         conn.execute(f"""DELETE FROM games
